adaptions for determinitic run,  speed: binomial distribution instead plant wise, missing amount in reproduction,

+++ src/module_interaction/interaction.cpp
@@ -45,7 +45,7 @@ void INTERACTION::calculateCumulativeLeafAreaIndexAcrossHeightLayers(UTILS utils
     /// go through all living plants in the community and add their leaf area to the respective height layers
     for (int cohortindex = 0; cohortindex < community.totalNumberOfCohortsInCommunity; cohortindex++)
     {
-        int plantAmount = community.allPlants.at(cohortindex)->amount;
+        double plantAmount = community.allPlants.at(cohortindex)->amount;
         int pft = community.allPlants.at(cohortindex)->pft;
         double plantHeight = community.allPlants.at(cohortindex)->height;
         double plantArea = community.allPlants.at(cohortindex)->coveredArea;
@@ -113,7 +113,7 @@ void INTERACTION::calculateLightAvailabilityForPlants(UTILS utils, COMMUNITY &co
 
     for (int cohortindex = 0; cohortindex < community.totalNumberOfCohortsInCommunity; cohortindex++)
     {
-        int plantAmount = community.allPlants.at(cohortindex)->amount;
+        double plantAmount = community.allPlants.at(cohortindex)->amount;
         int pft = community.allPlants.at(cohortindex)->pft;
         double plantHeight = community.allPlants.at(cohortindex)->height;
